requirements.txt
groq>=0.4.0
python-docx>=0.8.11
pypdf>=3.15.1
tomli>=2.0.1
python-dotenv>=1.0.0
rich>=13.4.2
typer>=0.9.0
pydantic>=2.4.2
loguru>=0.7.0 